+++ cocos/renderer/ccShader_PositionColor.frag
@@ -26,7 +26,8 @@
 const char* ccPositionColor_frag = R"(
 
 #ifdef GL_ES
-precision lowp float;
+// STEVE: precision lowp float;
+precision highp float;
 #endif
 
 varying vec4 v_fragmentColor;
